requirements.txt
asgiref==3.9.1
diff-match-patch==20241021
Django==5.2.4
django-auditlog==3.2.1
django-import-export==4.3.9
et_xmlfile==2.0.0
openpyxl==3.1.5
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
